moving functions that generate cm_info data structure to pynestml/utils.
Now they can be used by coco and the code generator separately in order
not to exchange data between coco and code generator.
